fix(396/B3): a2r stdlib 前导恢复裸 a2r_std 形式——修复 Plan 405 引入的双重回归
- Plan 405 以'a2r-std 独立 crate 不存在'为由把前导改为 auto_lang::a2r_std 限定形式，
  实际 crates/a2r-std 存在且是外部消费者（auto-ai 转译 crate）的声明依赖——限定形式
  使其 12 处 unresolved import，并使 48 个 golden 头部失效（此前误判为'环境差异'）
- rust.rs 恢复裸 use a2r_std; + 9 个 cookbook/concurrency golden 同步 Plan 399 的
  int→i64 决策（纯 i32→i64 差异，逐个校验后更新）
- api_gen 新增 qualify_a2r_std 后处理：生成的后端 Cargo.toml 依赖 auto-lang（非独立
  a2r-std），在 db.rs/api.rs 写盘前重限定为 auto_lang:: 形式——Plan 405 后端不受影响
- 验证：a2r golden 339/339 全绿（48 失败清零）；auto-ai retranspile 的 auto_lang 类
  错误 12→0

+++ crates/auto-lang/test/cookbook/concurrency/001_rayon_any_all/rayon_any_all.expected.rs
@@ -2,7 +2,7 @@
 
 use rayon::prelude::*;
 fn main() {
-    let v: Vec<i32> = vec![1, 2, 3, 4, 5];
+    let v: Vec<i64> = vec![1, 2, 3, 4, 5];
     let has_even = v.par_iter().any(|x| x % 2 == 0);
     let all_positive = v.par_iter().all(|x| x > 0);
     assert!(has_even == true);
