+++ crates/auto-lang/test/cookbook/concurrency/007_crossbeam_complex/crossbeam_complex.expected.rs
@@ -2,7 +2,7 @@
 
 fn main() {
     let messages: Vec<String> = vec!["hello".to_string(), "world".to_string()];
-    let mut msg_count: i32 = 0;
+    let mut msg_count: i64 = 0;
     for msg in &messages {
         msg_count += 1;
         if msg == "world" {
